separate aws/aks methods/functions to aws/aks go files
