+++ cloud/base.go
@@ -5,12 +5,9 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/sirupsen/logrus"
 	"net/http"
-	azureCluster "github.com/banzaicloud/azure-aks-client/cluster"
-	azureClient "github.com/banzaicloud/azure-aks-client/client"
 	"fmt"
 	"bytes"
 	"reflect"
-	"github.com/kris-nova/kubicorn/apis/cluster"
 	"errors"
 )
 
@@ -56,6 +53,18 @@ type ClusterSimple struct {
 	Azure            AzureSimple
 }
 
+type ClusterRepresentation struct {
+	Id        uint        `json:"id"`
+	Name      string      `json:"name"`
+	CloudType string      `json:"cloud"`
+	*AmazonRepresentation `json:"amazon"`
+	*AzureRepresentation  `json:"azure"`
+}
+
+type AmazonRepresentation struct {
+	Ip string `json:"ip"`
+}
+
 // String method prints formatted update request fields
 func (r UpdateClusterRequest) String() string {
 	var buffer bytes.Buffer
@@ -77,17 +86,17 @@ func (r UpdateClusterRequest) String() string {
 }
 
 // DeleteFromDb deletes cluster from database
-func (cluster ClusterSimple) DeleteFromDb(c *gin.Context, db *gorm.DB, log *logrus.Logger) bool {
+func (cs ClusterSimple) DeleteFromDb(c *gin.Context, db *gorm.DB, log *logrus.Logger) bool {
 
 	log.Info("Delete from db")
 
-	if err := db.Delete(&cluster).Error; err != nil {
+	if err := db.Delete(&cs).Error; err != nil {
 		// delete failed
 		log.Warning("Can't delete cluster from database!", err)
 		SetResponseBodyJson(c, http.StatusBadRequest, gin.H{
 			JsonKeyStatus:     http.StatusBadRequest,
 			JsonKeyMessage:    "Can't delete cluster!",
-			JsonKeyResourceId: cluster.ID,
+			JsonKeyResourceId: cs.ID,
 			JsonKeyError:      err,
 		})
 		return false
@@ -101,130 +110,29 @@ func (ClusterSimple) TableName() string {
 }
 
 // String method prints formatted cluster fields
-func (cluster ClusterSimple) String() string {
+func (cs ClusterSimple) String() string {
 	var buffer bytes.Buffer
-	buffer.WriteString(fmt.Sprintf("Id: %d, Creation date: %s, Cloud: %s, NodeInstanceType: %s, ", cluster.ID, cluster.CreatedAt, cluster.Cloud, cluster.NodeInstanceType))
-	if cluster.Cloud == Azure {
+	buffer.WriteString(fmt.Sprintf("Id: %d, Creation date: %s, Cloud: %s, NodeInstanceType: %s, ", cs.ID, cs.CreatedAt, cs.Cloud, cs.NodeInstanceType))
+	if cs.Cloud == Azure {
 		// Write AKS
 		buffer.WriteString(fmt.Sprintf("Agent count: %d, Agent name: %s, Kubernetes version: %s",
-			cluster.Azure.AgentCount,
-			cluster.Azure.AgentName,
-			cluster.Azure.KubernetesVersion))
-	} else if cluster.Cloud == Amazon {
+			cs.Azure.AgentCount,
+			cs.Azure.AgentName,
+			cs.Azure.KubernetesVersion))
+	} else if cs.Cloud == Amazon {
 		// Write AWS Master
 		buffer.WriteString(fmt.Sprintf("Master instance type: %s, Master image: %s",
-			cluster.Amazon.MasterInstanceType,
-			cluster.Amazon.MasterImage))
+			cs.Amazon.MasterInstanceType,
+			cs.Amazon.MasterImage))
 		// Write AWS Node
 		buffer.WriteString(fmt.Sprintf("Spot price: %s, Min count: %d, Max count: %d, Node image: %s",
-			cluster.Amazon.NodeSpotPrice,
-			cluster.Amazon.NodeMinCount,
-			cluster.Amazon.NodeMaxCount,
-			cluster.Amazon.NodeImage))
+			cs.Amazon.NodeSpotPrice,
+			cs.Amazon.NodeMinCount,
+			cs.Amazon.NodeMaxCount,
+			cs.Amazon.NodeImage))
 	}
 
 	return buffer.String()
-}
-
-// CreateClusterAmazon creates amazon cluster in cloud
-func (request CreateClusterRequest) CreateClusterAmazon(c *gin.Context, db *gorm.DB, log *logrus.Logger) (bool, *cluster.Cluster) {
-
-	cluster2Db := ClusterSimple{
-		Name:             request.Name,
-		Location:         request.Location,
-		NodeInstanceType: request.NodeInstanceType,
-		Cloud:            request.Cloud,
-		Amazon: AmazonClusterSimple{
-			NodeSpotPrice:      request.Properties.CreateClusterAmazon.Node.SpotPrice,
-			NodeMinCount:       request.Properties.CreateClusterAmazon.Node.MinCount,
-			NodeMaxCount:       request.Properties.CreateClusterAmazon.Node.MaxCount,
-			NodeImage:          request.Properties.CreateClusterAmazon.Node.Image,
-			MasterInstanceType: request.Properties.CreateClusterAmazon.Master.InstanceType,
-			MasterImage:        request.Properties.CreateClusterAmazon.Master.Image,
-		},
-	}
-
-	// create aws cluster
-	if createdCluster, err := CreateCluster(cluster2Db); err != nil {
-		// creation failed
-		log.Info("Cluster creation failed!", err)
-		SetResponseBodyJson(c, http.StatusBadRequest, gin.H{
-			JsonKeyMessage: "Could not launch cluster!",
-			JsonKeyName:    cluster2Db.Name,
-			JsonKeyError:   err,
-		})
-		return false, nil
-	} else {
-		// cluster creation success
-		log.Info("Cluster created successfully!")
-
-		// save db
-		if err := db.Save(&cluster2Db).Error; err != nil {
-			DbSaveFailed(c, log, err, cluster2Db.Name)
-			return false, nil
-		}
-
-		SetResponseBodyJson(c, http.StatusCreated, gin.H{
-			JsonKeyStatus:     http.StatusCreated,
-			JsonKeyMessage:    "Cluster created successfully!",
-			JsonKeyResourceId: cluster2Db.ID,
-			JsonKeyName:       cluster2Db.Name,
-			JsonKeyIp:         createdCluster.KubernetesAPI.Endpoint,
-		})
-
-		return true, createdCluster
-	}
-
-}
-
-// updateClusterAzureInCloud updates azure cluster in cloud
-func (r UpdateClusterRequest) updateClusterAzureInCloud(c *gin.Context, db *gorm.DB, log *logrus.Logger, preCluster ClusterSimple) bool {
-
-	cluster2Db := ClusterSimple{
-		Model:            preCluster.Model,
-		Name:             preCluster.Name,
-		Location:         preCluster.Location,
-		NodeInstanceType: preCluster.NodeInstanceType,
-		Cloud:            r.Cloud,
-		Azure: AzureSimple{
-			ResourceGroup:     preCluster.Azure.ResourceGroup,
-			AgentCount:        r.UpdateClusterAzure.AgentCount,
-			AgentName:         preCluster.Azure.AgentName,
-			KubernetesVersion: preCluster.Azure.KubernetesVersion,
-		},
-	}
-
-	ccr := azureCluster.CreateClusterRequest{
-		Name:              cluster2Db.Name,
-		Location:          cluster2Db.Location,
-		VMSize:            cluster2Db.NodeInstanceType,
-		ResourceGroup:     cluster2Db.Azure.ResourceGroup,
-		AgentCount:        cluster2Db.Azure.AgentCount,
-		AgentName:         cluster2Db.Azure.AgentName,
-		KubernetesVersion: cluster2Db.Azure.KubernetesVersion,
-	}
-
-	res, err := azureClient.CreateUpdateCluster(ccr)
-	if err != nil {
-		log.Info("Cluster update failed!", err.Message)
-		SetResponseBodyJson(c, err.StatusCode, gin.H{
-			JsonKeyStatus:  err.StatusCode,
-			JsonKeyMessage: err.Message,
-		})
-		return false
-	} else {
-		log.Info("Cluster update success")
-		// updateDb
-		if updateClusterInDb(c, db, log, cluster2Db) {
-			// success update
-			SetResponseBodyJson(c, res.StatusCode, res.Value)
-			return true
-		} else {
-			return false
-		}
-
-	}
-
 }
 
 func updateClusterInDb(c *gin.Context, db *gorm.DB, log *logrus.Logger, cluster ClusterSimple) bool {
@@ -235,62 +143,15 @@ func updateClusterInDb(c *gin.Context, db *gorm.DB, log *logrus.Logger, cluster 
 	return true
 }
 
-// updateClusterAmazonInCloud updates amazon cluster in cloud
-func (r UpdateClusterRequest) updateClusterAmazonInCloud(c *gin.Context, db *gorm.DB, log *logrus.Logger, preCluster ClusterSimple) bool {
-
-	cluster2Db := ClusterSimple{
-		Model:            preCluster.Model,
-		Name:             preCluster.Name,
-		Location:         preCluster.Location,
-		NodeInstanceType: preCluster.NodeInstanceType,
-		Cloud:            r.Cloud,
-		Amazon: AmazonClusterSimple{
-			NodeSpotPrice:      preCluster.Amazon.NodeSpotPrice,
-			NodeMinCount:       r.UpdateClusterAmazon.MinCount,
-			NodeMaxCount:       r.UpdateClusterAmazon.MaxCount,
-			NodeImage:          preCluster.Amazon.NodeImage,
-			MasterInstanceType: preCluster.Amazon.MasterInstanceType,
-			MasterImage:        preCluster.Amazon.MasterImage,
-		},
-	}
-
-	if _, err := UpdateClusterAws(cluster2Db); err != nil {
-		log.Warning("Can't update cluster in the cloud!", err)
-
-		SetResponseBodyJson(c, http.StatusBadRequest, gin.H{
-			JsonKeyStatus:     http.StatusBadRequest,
-			JsonKeyMessage:    "Can't update cluster in the cloud!",
-			JsonKeyResourceId: cluster2Db.ID,
-			JsonKeyError:      err,
-		})
-
-		return false
-	} else {
-		log.Info("Cluster updated in the cloud!")
-		if updateClusterInDb(c, db, log, cluster2Db) {
-			SetResponseBodyJson(c, http.StatusCreated, gin.H{
-				JsonKeyStatus:     http.StatusCreated,
-				JsonKeyMessage:    "Cluster updated successfully!",
-				JsonKeyResourceId: cluster2Db.ID,
-			})
-
-			return true
-		}
-
-		return false
-	}
-
-}
-
 // UpdateClusterInCloud updates cluster in cloud
 // The request's cloud field decided which type of cloud will be called
 func (r *UpdateClusterRequest) UpdateClusterInCloud(c *gin.Context, db *gorm.DB, log *logrus.Logger, preCluster ClusterSimple) bool {
 
 	switch r.Cloud {
 	case Amazon:
-		return r.updateClusterAmazonInCloud(c, db, log, preCluster)
+		return r.UpdateClusterAmazonInCloud(c, db, log, preCluster)
 	case Azure:
-		return r.updateClusterAzureInCloud(c, db, log, preCluster)
+		return r.UpdateClusterAzureInCloud(c, db, log, preCluster)
 	default:
 		return false
 	}
@@ -306,118 +167,12 @@ func (r *UpdateClusterRequest) Validate(log *logrus.Logger, defaultValue Cluster
 		return r.ValidateAmazonRequest(log, defaultValue)
 	case Azure:
 		// azure validate
-		return r.validateAzureRequest(log, defaultValue)
+		return r.ValidateAzureRequest(log, defaultValue)
 	default:
 		// not supported cloud type
 		return false, "Not supported cloud type."
 	}
 
-}
-
-// ValidateAmazonRequest validates the update request (only amazon part). If any of the fields is missing, the method fills
-// with stored data.
-func (r *UpdateClusterRequest) ValidateAmazonRequest(log *logrus.Logger, defaultValue ClusterSimple) (bool, string) {
-
-	// reset azure fields
-	r.UpdateClusterAzure = nil
-
-	defAmazonNode := &UpdateAmazonNode{
-		MinCount: defaultValue.Amazon.NodeMinCount,
-		MaxCount: defaultValue.Amazon.NodeMaxCount,
-	}
-
-	// ---- [ Amazon field check ] ---- //
-	if r.UpdateClusterAmazon == nil {
-		log.Info("'amazon' field is empty, Load it from stored data.")
-		r.UpdateClusterAmazon = &UpdateClusterAmazon{
-			UpdateAmazonNode: defAmazonNode,
-		}
-	}
-
-	// ---- [ Node check ] ---- //
-	if r.UpdateAmazonNode == nil {
-		log.Info("'node' field is empty. Fill from stored data")
-		r.UpdateAmazonNode = defAmazonNode
-	}
-
-	// ---- [ Node min count check ] ---- //
-	if r.UpdateAmazonNode.MinCount == 0 {
-		defMinCount := defaultValue.Amazon.NodeMinCount
-		log.Info("Node minCount set to default value: ", defMinCount)
-		r.UpdateAmazonNode.MinCount = defMinCount
-	}
-
-	// ---- [ Node max count check ] ---- //
-	if r.UpdateAmazonNode.MaxCount == 0 {
-		defMaxCount := defaultValue.Amazon.NodeMaxCount
-		log.Info("Node maxCount set to default value: ", defMaxCount)
-		r.UpdateAmazonNode.MaxCount = defMaxCount
-	}
-
-	// ---- [ Node max count > min count check ] ---- //
-	if r.UpdateAmazonNode.MaxCount < r.UpdateAmazonNode.MinCount {
-		log.Info("Node maxCount is lower than minCount")
-		return false, "maxCount must be greater than mintCount"
-	}
-
-	// create update request struct with the stored data to check equality
-	preCl := &UpdateClusterRequest{
-		Cloud: defaultValue.Cloud,
-		UpdateProperties: UpdateProperties{
-			UpdateClusterAmazon: &UpdateClusterAmazon{
-				UpdateAmazonNode: defAmazonNode,
-			},
-		},
-	}
-
-	// check equality
-	return isUpdateEqualsWithStoredCluster(r, preCl, log)
-}
-
-// ValidateAzureRequest validates the update request (only azure part). If any of the fields is missing, the method fills
-// with stored data.
-func (r *UpdateClusterRequest) validateAzureRequest(log *logrus.Logger, defaultValue ClusterSimple) (bool, string) {
-
-	// reset field amazon fields
-	r.UpdateClusterAmazon = nil
-
-	defAzureNode := &UpdateAzureNode{
-		AgentCount: defaultValue.Azure.AgentCount,
-	}
-
-	// ---- [ Azure field check ] ---- //
-	if r.UpdateClusterAzure == nil {
-		log.Info("'azure' field is empty, Load it from stored data.")
-		r.UpdateClusterAzure = &UpdateClusterAzure{
-			UpdateAzureNode: defAzureNode,
-		}
-	}
-
-	// ---- [ Node check ] ---- //
-	if r.UpdateClusterAzure.UpdateAzureNode == nil {
-		log.Info("'node' field is empty. Load it from stored data.")
-		r.UpdateClusterAzure.UpdateAzureNode = defAzureNode
-	}
-
-	// ---- [ Node - Agent count check] ---- //
-	if r.UpdateClusterAzure.AgentCount == 0 {
-		def := defaultValue.Azure.AgentCount
-		log.Info("Node agentCount set to default value: ", def)
-		r.UpdateClusterAzure.AgentCount = def
-	}
-
-	// create update request struct with the stored data to check equality
-	preCl := &UpdateClusterRequest{
-		Cloud: defaultValue.Cloud,
-		UpdateProperties: UpdateProperties{
-			UpdateClusterAzure: &UpdateClusterAzure{
-				UpdateAzureNode: defAzureNode,
-			},
-		},
-	}
-
-	// check equality
-	return isUpdateEqualsWithStoredCluster(r, preCl, log)
 }
 
 // isUpdateEqualsWithStoredCluster compares x and y interfaces with deep equal
@@ -430,56 +185,6 @@ func isUpdateEqualsWithStoredCluster(x interface{}, y interface{}, log *logrus.L
 	return true, ""
 }
 
-// CreateClusterAzure creates azure cluster in the cloud
-func (request CreateClusterRequest) CreateClusterAzure(c *gin.Context, db *gorm.DB, log *logrus.Logger) bool {
-
-	cluster2Db := ClusterSimple{
-		Name:             request.Name,
-		Location:         request.Location,
-		NodeInstanceType: request.NodeInstanceType,
-		Cloud:            request.Cloud,
-		Azure: AzureSimple{
-			ResourceGroup:     request.Properties.CreateClusterAzure.Node.ResourceGroup,
-			AgentCount:        request.Properties.CreateClusterAzure.Node.AgentCount,
-			AgentName:         request.Properties.CreateClusterAzure.Node.AgentName,
-			KubernetesVersion: request.Properties.CreateClusterAzure.Node.KubernetesVersion,
-		},
-	}
-
-	r := azureCluster.CreateClusterRequest{
-		Name:              cluster2Db.Name,
-		Location:          cluster2Db.Location,
-		VMSize:            cluster2Db.NodeInstanceType,
-		ResourceGroup:     cluster2Db.Azure.ResourceGroup,
-		AgentCount:        cluster2Db.Azure.AgentCount,
-		AgentName:         cluster2Db.Azure.AgentName,
-		KubernetesVersion: cluster2Db.Azure.KubernetesVersion,
-	}
-
-	// call creation
-	res, err := azureClient.CreateUpdateCluster(r)
-	if err != nil {
-		// creation failed
-		log.Info("Cluster creation failed!", err.Message)
-		SetResponseBodyJson(c, err.StatusCode, gin.H{
-			JsonKeyStatus:  err.StatusCode,
-			JsonKeyMessage: err.Message,
-		})
-		return false
-	} else {
-		// creation success
-		log.Info("Cluster created successfully!")
-		if err := db.Save(&cluster2Db).Error; err != nil {
-			DbSaveFailed(c, log, err, cluster2Db.Name)
-			return false
-		}
-
-		SetResponseBodyJson(c, res.StatusCode, res.Value)
-		return true
-	}
-
-}
-
 // DbSaveFailed sends DB operation failed message back
 func DbSaveFailed(c *gin.Context, log *logrus.Logger, err error, clusterName string) {
 	log.Warning("Can't persist cluster into the database!", err)
@@ -490,92 +195,6 @@ func DbSaveFailed(c *gin.Context, log *logrus.Logger, err error, clusterName str
 		JsonKeyName:    clusterName,
 		JsonKeyError:   err,
 	})
-}
-
-func (cluster *ClusterSimple) GetAzureClusterStatus(c *gin.Context, db *gorm.DB, log *logrus.Logger) {
-	// load azure props from db
-	db.Where(AzureSimple{ClusterSimpleId: cluster.ID}).First(&cluster.Azure)
-	resp, err := azureClient.GetCluster(cluster.Name, cluster.Azure.ResourceGroup)
-	if err != nil {
-		log.Info("Error during get cluster info: ", err.Message)
-		SetResponseBodyJson(c, http.StatusInternalServerError, gin.H{
-			JsonKeyStatus:  http.StatusInternalServerError,
-			JsonKeyMessage: err.Message,
-		})
-	} else {
-		log.Info("Get cluster success")
-		stage := resp.Value.Properties.ProvisioningState
-		var msg string
-		var code int
-		if stage == "Succeeded" {
-			msg = "Cluster available"
-			code = http.StatusOK
-		} else {
-			msg = "Cluster not ready yet"
-			code = http.StatusNoContent
-		}
-		SetResponseBodyJson(c, code, gin.H{
-			JsonKeyStatus:  code,
-			JsonKeyMessage: msg,
-		})
-	}
-}
-
-func (cluster *ClusterSimple) GetAmazonClusterStatus(c *gin.Context, log *logrus.Logger) {
-	cl, err := cluster.GetClusterWithDbCluster(c, log)
-	if err != nil {
-		log.Info("Error during read cluster from db")
-		return
-	}
-	isAvailable, _ := IsKubernetesClusterAvailable(cl)
-	if isAvailable {
-		msg := "Kubernetes cluster available"
-		log.Info(msg)
-		SetResponseBodyJson(c, http.StatusOK, gin.H{
-			JsonKeyStatus:  http.StatusOK,
-			JsonKeyMessage: msg,
-		})
-	} else {
-		msg := "Kubernetes cluster not ready yet"
-		log.Info(msg)
-		SetResponseBodyJson(c, http.StatusNoContent, gin.H{
-			JsonKeyStatus:  http.StatusNoContent,
-			JsonKeyMessage: msg,
-		})
-	}
-
-}
-
-func (cluster *ClusterSimple) GetClusterWithDbCluster(c *gin.Context, log *logrus.Logger) (*cluster.Cluster, error) {
-	cl, err := cluster.GetKubicornCluster(log)
-	if err != nil {
-		errorMsg := fmt.Sprintf("Error read cluster: %s", err)
-		SetResponseBodyJson(c, http.StatusNotFound, gin.H{
-			JsonKeyStatus:  http.StatusNotFound,
-			JsonKeyMessage: errorMsg,
-		})
-		return nil, err
-	}
-	return cl, nil
-}
-
-// GetCluster based on ClusterSimple object
-// This will read the persisted Kubicorn cluster format
-func (cluster *ClusterSimple) GetKubicornCluster(log *logrus.Logger) (*cluster.Cluster, error) {
-	clust, err := ReadCluster(*cluster)
-	if err != nil {
-		return nil, err
-	}
-	log.Info("Cluster read successful")
-	return clust, nil
-}
-
-func GetCluster(c *gin.Context, db *gorm.DB, log *logrus.Logger) (*cluster.Cluster, error) {
-	cl, err := GetClusterFromDB(c, db)
-	if err != nil {
-		return nil, err
-	}
-	return cl.GetClusterWithDbCluster(c, log)
 }
 
 // GetCluster from database
@@ -608,4 +227,74 @@ func GetClusterSimple(c *gin.Context, db *gorm.DB) (*ClusterSimple, error) {
 		return nil, err
 	}
 	return cl, nil
+}
+
+func (cs *ClusterSimple) DeleteCluster(c *gin.Context, db *gorm.DB, log *logrus.Logger) bool {
+
+	clusterType := cs.Cloud
+	log.Info("Cluster type is ", clusterType)
+
+	switch clusterType {
+	case Amazon:
+		// create amazon cs
+		return cs.DeleteAmazonCluster(c, db, log)
+	case Azure:
+		// delete azure cs
+		return cs.DeleteAzureCluster(c, db, log)
+	default:
+		SendNotSupportedCloudResponse(c)
+		return false
+	}
+
+}
+
+// SendNotSupportedCloudResponse sends Not-supported-cloud-type error message back
+func SendNotSupportedCloudResponse(c *gin.Context) {
+	msg := "Not supported cloud type. Please use one of the following: " + Amazon + ", " + Azure + "."
+	SetResponseBodyJson(c, http.StatusBadRequest, gin.H{
+		JsonKeyStatus:  http.StatusBadRequest,
+		JsonKeyMessage: msg,
+	})
+}
+
+func (cs *ClusterSimple) GetClusterRepresentation(db *gorm.DB, log *logrus.Logger) *ClusterRepresentation {
+	cloudType := cs.Cloud
+	log.Info("Cloud type is ", cloudType)
+	switch cloudType {
+	case Amazon:
+		return cs.ReadClusterAmazon(log)
+		break
+	case Azure:
+		db.Where(AzureSimple{ClusterSimpleId: cs.ID}).First(&cs.Azure)
+		return cs.ReadClusterAzure(log)
+		break
+	default:
+		log.Info("Not supported cloud type")
+		break
+	}
+	return nil
+}
+
+func (cs *ClusterSimple) FetchClusterInfo(c *gin.Context, db *gorm.DB, log *logrus.Logger) {
+
+	cloudType := cs.Cloud
+	log.Info("Cloud type is ", cloudType)
+
+	switch cloudType {
+	case Amazon:
+		cs.GetClusterInfoAmazon(c, log)
+		break
+	case Azure:
+		// set azure props
+		db.Where(AzureSimple{ClusterSimpleId: cs.ID}).First(&cs.Azure)
+		cs.GetClusterInfoAzure(c, log)
+		break
+	default:
+		// wrong cloud type
+		SetResponseBodyJson(c, http.StatusInternalServerError, gin.H{
+			JsonKeyStatus:  http.StatusInternalServerError,
+			JsonKeyMessage: "Not supported cloud type.",
+		})
+		break
+	}
 }
