+++ utils/utils.go
@@ -1,6 +1,9 @@
 package utils
 
-import "os"
+import (
+	"os"
+	"strconv"
+)
 
 //GetEnv retrieves ENV variable, fallback if not set
 func GetEnv(envKey, defaultValue string) string {
@@ -15,4 +18,13 @@ func GetEnv(envKey, defaultValue string) string {
 func GetHomeDir() string {
 	//Linux
 	return os.Getenv("HOME")
+}
+
+// convertString2Uint converts a string to uint
+func ConvertString2Uint(s string) uint {
+	i, err := strconv.ParseInt(s, 10, 32)
+	if err != nil {
+		panic(err)
+	}
+	return uint(i)
 }
